fix(app-shell): coercing values for stringified array with double escaped quotes
