+++ packages/application-shell/src/components/application-shell/fixtures/coerce-values.json
@@ -1,0 +1,10 @@
+{
+  "empty": "",
+  "bool": "true",
+  "array": "[\"1\", \"2\"]",
+  "arrayWithDoubleEscapedQuotes": "[\\\"1\\\", \\\"2\\\"]",
+  "emptyArray": "[]",
+  "null": "null",
+  "number": "1",
+  "string": "foo"
+}
